refactor: move existing bolt-versions task to new api-tasks folder in @bolt/build-tools
